Assignment #2 - Part 3:  Sign In
No Optionals:

Sign In
Add custom nav bar as a UIImageView, form background, text fields,
text, and Sign In button.
Tapping the Sign In button with either a blank email or blank password
should show an error alert. Using UIAlertController
Tapping the Sign In button with a populated email and password field
should display a loading alert for 2 seconds. After the 2 seconds are
up, you should do one of the following (Calling a Method After Delay):
Display an error for incorrect email/password.
Go to the tutorial screens.